move non prisma files into src folder with web subfolder to represent frontend specific
