+++ src/utils/sportradar-access-functions.ts
