+++ src/web/components/homepage-grid/every-player-container.tsx
